examples/soccer/requirements.txt
gdown
google-cloud-speech
google-cloud-texttospeech
google-cloud-translate
moviepy
numpy
opencv-python
sounddevice
soundfile
supervision
ultralytics
